+++ modules/platform/src/main/kotlin/dev/sumin/skeleton/common/audit/BaseAuditTimestamps.kt
@@ -1,0 +1,9 @@
+package dev.sumin.skeleton.common.audit
+
+import java.time.Instant
+
+interface BaseAuditTimestamps {
+    val createdAt: Instant
+    val updatedAt: Instant
+    val deletedAt: Instant?
+}